requirements.txt
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
pytube==15.0.0
flask==3.0.0
transformers==4.37.2
torch
sentencepiece
Markdown==3.5.1 